Refactor payment service injection and update service annotations for Paypal and Stripe

+++ src/main/java/com/elkabani/firstspringboot/repository/CategoryRepository.java
@@ -2,8 +2,9 @@ package com.elkabani.firstspringboot.repository;
 
 import com.elkabani.firstspringboot.model.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import java.util.Optional;
 
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
     Optional<Category> findByName(String name);
 }
